IOS-7660: If initialized with a folder other than root can't browse full account
Created a new class method to get the full array of parent folder view
controllers that would be needed to allow a user to start at some specified
folder and browse all the way back to the root (All Files).

Also changed around the designated initializer of the folder view controller.