Crash getting resources from bundle
Apple's documentation says that imageNamed:inBundle:compatibleWithTraitCollection:
is thread safe as of iOS 9. We are seeing crashes in both iOS 8 and 9 on calling
it, and it is always in process on another thread (or more) as well. It is possible
that Apple's iOS 9 fix does not make it thread safe in all scenarios.
Additionally, the NSBundle object returned will be the same most of the time,
from Apple's documentation: "This method allocates and initializes the returned
object if there is no existing NSBundle associated with fullPath, in which case
it returns the existing object."